fix(api-file-manager): use Long type in getPreSignedPostPayload query

+++ packages/api-file-manager-s3/src/plugins/graphqlFileStorageS3.ts
@@ -15,13 +15,13 @@ const plugin: GraphQLSchemaPlugin<FileManagerContext> = {
             input PreSignedPostPayloadInput {
                 name: String!
                 type: String!
-                size: Int!
+                size: Long!
             }
 
             type GetPreSignedPostPayloadResponseDataFile {
                 name: String
                 type: String
-                size: Int
+                size: Long
                 key: String
             }
